refactor: extract CET taxonomy module as pilot refactoring
This is the first phase of refactoring large CET asset files into
modular components as outlined in the asset refactoring plan.

**Changes:**

New modular structure in src/assets/cet/:
- utils.py: Shared utilities (Dagster shims, I/O functions, metrics serialization)
- taxonomy.py: Taxonomy asset, checks, and helper functions
- __init__.py: Package exports for backward compatibility

Updated src/assets/cet_assets.py:
- Import taxonomy module from .cet subpackage
- Remove duplicate definitions of taxonomy-related code
- Maintain backward compatibility for existing imports

**Impact:**
- Reduced cet_assets.py from 3,597 lines (pilot extraction only)
- Extracted 3 new focused modules (~200 lines each)
- All imports remain backward compatible
- No breaking changes to existing code

**Testing:**
- Verified Python syntax with py_compile
- Checked backward compatibility of imports in cet_pipeline_job.py
- Confirmed no breaking changes to existing code paths

**Next Steps:**
Per the refactoring plan, continue with:
- Week 1: Complete CET assets refactoring (classifications, training, analytics, validation, company, loading)
- Week 2: USPTO assets refactoring
- Week 3: Transition assets refactoring

Related: docs/architecture/asset-refactoring-plan.md
